requirements.txt
git+http://github.com/TeHikuMedia/reo-toolkit.git
requests
GitPython
nltk
pyyaml
pandas
